// src/app/Models/item.model.ts
/**
 * Dashboard APIs
 * Dashboard API.
 *
 * OpenAPI spec version: 1.0.0
 * Contact: daniel@delicio.io
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { CategoryDto } from './Cat_Product.model';


export class ItemDto {
  /**
   * GUID of the Item. This is optional input for POST operation i.e., creating item.
   */
  id?: string;
  /**
   * item Name
   */
  itemName?: string = '';
  /**
   * item Description
   */
  itemDescription?: string = '';
  /**
   * item Unit Price
   */
  itemUnitPrice?: number;
  /**
   * item SKU
   */
  itemSKU?: string = '';
  /**
   * Status. 0 - Inactive, 1 - Active
   */
  itemStatus?: number = 1;
  /**
   * Number of addONs associated with this product
   */
  addONCount?: number;
  addonOptionCount?: number;
  imageUrl?: string;
  itemType?: number = 0;//either a product or ad on?
  itemPriceMappings: PlatfromPricing[] = [];
  selected: boolean = true;//field for managing checkboxes in assigning items to cats
  posName: string;
  shortName: string;
  serviceAvailability: ItemServiceAvailability[] = []
}

export class ItemServiceAvailability {
  id:any = ''
  itemId:string = ''
  startTime: string = ''
  endTime: string = ''
  weekDay: string = ''
}

export class ItemDtoNew {
  /**
   * GUID of the Item. This is optional input for POST operation i.e., creating item.
   */
  id?: string;
  /**
   * item Name
   */
  itemName?: string = '';
  /**
   * item Description
   */
  itemDescription?: string = '';
  /**
   * item Unit Price
   */
  itemUnitPrice?: number;
  /**
   * item SKU
   */
  itemSKU?: string = '';
  /**
   * Status. 0 - Inactive, 1 - Active
   */
  itemStatus?: number = 1;
  /**
   * Number of addONs associated with this product
   */
  addONCount?: number;
  categoryCount?: number;
  imageUrl?: string;
  itemType?: number = 0;//either a product or ad on?
  itemPriceMappings: PlatfromPricing[] = [];
  selected: boolean = true;//field for managing checkboxes in assigning items to cats
  posName: string;
  shortName: string;

}


export class AddonOption {
  id: string;
  addonOptionName: string;
  addonOptionDescription: string;
  addonOptionPrice: number;
  status: boolean = false;
  addonOptionSKU: string;
  addonOptionStatus: number;
  selected: unknown;

}

export class AddonItemDetailDto {
  id: string;
  addonName: string;
  addonDescription: string;
  addonOptions: AddonOption[];
  minPermitted: number;
  maxPermitted: number
  enabled: boolean = false;//for storefront min max check
}

export class ItemDetailObject {
  id: string;
  itemName: string = "";
  itemDescription: string = "";
  itemUnitPrice: number = 0;
  itemSKU: string;
  itemStatus: number;
  posName: string;
  shortName: string;
  imageUrl: string;
  itemType?: number = 0;//either a product or ad on?
  categories: CategoryDto[] = [];
  addons: AddonItemDetailDto[] = [];
  itemPriceMappings: PlatfromPricing[] = [];
  specialInstructions: string = ''
  quantity: number = 1
  selected: boolean = true;
  serviceAvailability: ItemServiceAvailability[] = []
}

export class PlatfromPricing {
  platformType: number;
  platformItemPrice: number
}